Migrate mod compat screen and most mixins to common source set
