+++ neoforge/src/main/resources/inventorypause.neoforge.mixins.json
@@ -7,10 +7,8 @@
   "mixins": [
   ],
   "client": [
-    "IntegratedServerMixin",
-    "MinecraftMixin",
-    "KeyboardHandlerMixin",
-    "qol.DeathScreenMixin"
+    "KeyboardHandlerMixinNeo",
+    "MinecraftMixinNeo"
   ],
   "injectors": {
     "defaultRequire": 1
